feature: 최근 조회 상품 목록 구현. 상품 옵션 선택 후 주문 및 주문 정보 및 결제 정보 생성 구현.
refactor: 상품 목록 component로 리팩토링.
상품 가격 불러오는 코드 분리 및 리팩토링

+++ src/main/java/com/jay/shoppingmall/domain/item/item_stock/item_stock_history/ItemStockHistory.java
@@ -12,14 +12,13 @@ import java.time.LocalDateTime;
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-//가격이 변동될 때마다 무조건 새로 생성.
 public class ItemStockHistory {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     private Long id;
 
-    private Long stock;
+    private Integer stock;
 
     private LocalDateTime stockChangedDate;
